LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssTransformValue-toMatrix-expected.txt


FAIL CSSTransformValue.toMatrix() multiplies its component matrices assert_array_approx_equals: property 0, expected 10 +/- 0.000001, expected 10 but got 1
FAIL CSSTransformValue.toMatrix() respects is2D changes in its components assert_array_approx_equals: property 0, expected -6 +/- 0.000001, expected -6 but got 1
